+++ sample/src/androidTest/java/com/stepstone/stepper/sample/test/runner/TestButlerRunner.java
@@ -1,0 +1,25 @@
+package com.stepstone.stepper.sample.test.runner;
+
+import android.os.Bundle;
+import android.support.test.InstrumentationRegistry;
+import android.support.test.runner.AndroidJUnitRunner;
+
+import com.linkedin.android.testbutler.TestButler;
+
+/**
+ * @author Piotr Zawadzki
+ */
+public class TestButlerRunner extends AndroidJUnitRunner {
+
+    @Override
+    public void onStart() {
+        TestButler.setup(InstrumentationRegistry.getTargetContext());
+        super.onStart();
+    }
+
+    @Override
+    public void finish(int resultCode, Bundle results) {
+        TestButler.teardown(InstrumentationRegistry.getTargetContext());
+        super.finish(resultCode, results);
+    }
+}